LibWeb: Stop leaking visual context node types into unrelated headers
Two headers pulled in AccumulatedVisualContext.h for things that are not
visual context tree concerns. SVGPatternElement and SVGGraphicsElement
stored a pattern tile's content transform as a Painting::TransformData
whose only field ever read was the matrix, so they now hold the
Gfx::FloatMatrix4x4 directly. apply_css_transform_to_rect() maps a rect
through a box's own CSS transform without touching the tree; it lived in
the tree's header only by proximity and now sits with the other
Layout::Node geometry facades in BoxViews, which is what ScrollSnap, its
sole caller, already included.

This is preparation for shrinking AccumulatedVisualContext.h down to the
tree handle and index types once the tree is owned by Rust: nothing
outside the painting pipeline should need its node payload structs.
