+++ Libraries/LibWeb/Painting/BoxViews.h
@@ -76,6 +76,7 @@ WEB_API bool is_svg_paintable(Layout::Node const&);
 WEB_API bool is_svg_svg_paintable(Layout::Node const&);
 WEB_API bool is_svg_path_paintable(Layout::Node const&);
 
+WEB_API CSSPixelRect apply_css_transform_to_rect(Layout::Node const&, CSSPixelRect const&);
 WEB_API CSSPixelRect transform_rect_to_viewport(Layout::Node const&, CSSPixelRect const&, AccumulatedVisualContextTree::IncludeVisualViewportTransform = AccumulatedVisualContextTree::IncludeVisualViewportTransform::Yes);
 WEB_API Optional<CSSPixelPoint> transform_point_to_local(Layout::Node const&, CSSPixelPoint);
 WEB_API CSSPixelPoint inverse_transform_point(Layout::Node const&, CSSPixelPoint);
