+++ Libraries/LibWeb/SVG/SVGPatternElement.h
@@ -7,9 +7,9 @@
 #pragma once
 
 #include <LibGC/RootHashTable.h>
+#include <LibGfx/Matrix4x4.h>
 #include <LibWeb/Forward.h>
 #include <LibWeb/Layout/Node.h>
-#include <LibWeb/Painting/AccumulatedVisualContext.h>
 #include <LibWeb/SVG/AttributeParsing.h>
 #include <LibWeb/SVG/SVGAnimatedLength.h>
 #include <LibWeb/SVG/SVGElement.h>
@@ -57,7 +57,7 @@ public:
         Layout::NodeWithStyle const* pattern_layout_node;
         Gfx::FloatRect tile_rect;
         Gfx::FloatSize content_scale;
-        Painting::TransformData tile_content_transform;
+        Gfx::FloatMatrix4x4 tile_content_transform;
         Optional<Gfx::AffineTransform> device_pattern_transform;
     };
     Optional<PaintGeometry> resolve_paint_geometry(SVGPaintContext const&, double device_pixels_per_css_pixel, Layout::Node const& target_layout_node) const;
